fix #2931

+++ becpg-core/src/main/java/fr/becpg/repo/entity/datalist/policy/SortableListPolicy.java
@@ -84,7 +84,7 @@ public class SortableListPolicy extends AbstractBeCPGPolicy
 
 	@Override
 	public void onUpdateProperties(NodeRef nodeRef, Map<QName, Serializable> before, Map<QName, Serializable> after) {
-		if (policyBehaviourFilter.isEnabled(BeCPGModel.ASPECT_DEPTH_LEVEL)) {
+		if (policyBehaviourFilter.isEnabled(BeCPGModel.ASPECT_SORTABLE_LIST)) {
 			// createNode
 			if (before.isEmpty()) {
 				// nothing to do, work is done in addAspect, otherwise it
@@ -147,7 +147,7 @@ public class SortableListPolicy extends AbstractBeCPGPolicy
 	@Override
 	public void onAddAspect(NodeRef nodeRef, QName aspect) {
 
-		if (policyBehaviourFilter.isEnabled(BeCPGModel.ASPECT_DEPTH_LEVEL)) {
+		if (policyBehaviourFilter.isEnabled(BeCPGModel.ASPECT_DEPTH_LEVEL) && policyBehaviourFilter.isEnabled(BeCPGModel.ASPECT_SORTABLE_LIST)) {
 			// try to avoid to do two times the work, otherwise it duplicates
 			// nodeRef in lucene index !!!
 			if (nodeService.exists(nodeRef)) {
@@ -219,14 +219,16 @@ public class SortableListPolicy extends AbstractBeCPGPolicy
 	public void onDeleteNode(ChildAssociationRef childRef, boolean isNodeArchived) {
 
 		logger.debug("SortableListPolicy.onDeleteNode");
-
-		// if folder is deleted, all children are
-		if (nodeService.exists(childRef.getParentRef())) {
-			try {
-				policyBehaviourFilter.disableBehaviour(BeCPGModel.ASPECT_DEPTH_LEVEL);
-				dataListSortService.deleteChildrens(childRef.getParentRef(), childRef.getChildRef());
-			} finally {
-				policyBehaviourFilter.enableBehaviour(BeCPGModel.ASPECT_DEPTH_LEVEL);
+		if (policyBehaviourFilter.isEnabled(BeCPGModel.ASPECT_SORTABLE_LIST)) {
+	
+			// if folder is deleted, all children are
+			if (nodeService.exists(childRef.getParentRef())) {
+				try {
+					policyBehaviourFilter.disableBehaviour(BeCPGModel.ASPECT_DEPTH_LEVEL);
+					dataListSortService.deleteChildrens(childRef.getParentRef(), childRef.getChildRef());
+				} finally {
+					policyBehaviourFilter.enableBehaviour(BeCPGModel.ASPECT_DEPTH_LEVEL);
+				}
 			}
 		}
 	}
